refactor(builtins): separate RuleCollector from Rule via rules/collectors/ subdir
Move ResourceActionPermissionRuleCollector and ResourceActionScopeRuleCollector
(plus their tests) from packages/builtins/src/rules/ into
packages/builtins/src/rules/collectors/. Rule implementations stay at the
rules/ root; RuleCollector implementations now live in the dedicated
subdirectory. Exported symbol surface, class names, and module registry keys
are unchanged. Internal imports, barrel exports, and module wiring updated.

Spec: docs/superpowers/specs/2026-04-17-rule-spec-optimization-and-extension-guide-design.md
(in the agentscope workspace).
