+++ packages/builtins/src/rules/collectors/ResourceActionPermissionRuleCollector.mts
@@ -1,5 +1,5 @@
 import type { CollectorContext, Rule, RuleCollector } from "@o3co/auth.policy-verifier.core";
-import { HasPermission } from "./HasPermission.mjs";
+import { HasPermission } from "../HasPermission.mjs";
 
 export class ResourceActionPermissionRuleCollector implements RuleCollector {
 	async collect(context: CollectorContext): Promise<Rule[]> {
